src/test/java/ru/netology/stats/StatsServiceTest.java
package ru.netology.stats;

import org.junit.jupiter.api.Assertions;
import org.junit.jupiter.api.Test;

public class StatsServiceTest {

    @Test
    public void TotalSalesTest() {
        StatsService service = new StatsService();
        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 180;
        int actual = service.totalSalesByMonth(sales);

        Assertions.assertEquals(expected, actual);
    }

    @Test
    public void MidSalesTest() {
        StatsService service = new StatsService();
        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 15;
        int actual = service.midSalesByMonth(sales);

        Assertions.assertEquals(expected, actual);
    }

    @Test
    public void MaxSalesTest() {
        StatsService service = new StatsService();
        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 8;
        int actual = service.maxSalesByMonth(sales);

        Assertions.assertEquals(expected, actual);

    }

    @Test
    public void MinSalesTest() {
        StatsService service = new StatsService();
        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 9;
        int actual = service.minSalesByMonth(sales);

        Assertions.assertEquals(expected, actual);

    }

    @Test
    public void LowMidSalesTest() {
        StatsService service = new StatsService();
        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 5;
        int actual = service.lowMidSalesByMonth(sales);

        Assertions.assertEquals(expected, actual);

    }

    @Test
    public void HiMidSalesTest() {
        StatsService service = new StatsService();
        int[] sales = {8, 15, 13, 15, 17, 20, 19, 20, 7, 14, 14, 18};
        int expected = 5;
        int actual = service.lowMidSalesByMonth(sales);

        Assertions.assertEquals(expected, actual);

    }
}